Update README.md
With a lot of undergraduates graduating over the years, the business problem is around them and their pay scales, when they are in industry. The business problems that we are trying to tackle are: How much should a post graduate request for a salary  based on the type of school that they attended (liberal arts, engineering) ? How much do post graduates negotiate for a salary based on the major that they pursued? How much can a post graduate be paid based on where their school was located? Predicting the salaries of graduates based on majors, location and type of school combined. To be precise, the data analytics problem is to see how the region, type of school and major all affect the salary of a bachelor's student post graduation.